refactor: 점수를 더하는 역할 분리

+++ src/main/java/domain/Player.java
@@ -25,7 +25,7 @@ public class Player {
     }
 
     public int score() {
-        return hand.score();
+        return hand.score().toInt();
     }
 
     public boolean isBust() {
